some refactoring and re-exporting sui deps in sui-client
