+++ crates/sui/src/sui_commands.rs
@@ -14,7 +14,7 @@ use std::io::{stderr, stdout, Write};
 use std::num::NonZeroUsize;
 use std::path::{Path, PathBuf};
 use std::{fs, io};
-use sui_client::keystore::{KeystoreType, SuiKeystore};
+use sui_client::crypto::{KeystoreType, SuiKeystore};
 use sui_client::SuiRpcClient;
 use sui_config::{builder::ConfigBuilder, NetworkConfig, SUI_DEV_NET_URL, SUI_KEYSTORE_FILENAME};
 use sui_config::{genesis_config::GenesisConfig, SUI_GENESIS_FILENAME};
