+++ sdk/rust/examples/tic-tac-toe/README.md
@@ -1,0 +1,7 @@
+# Rust SDK example : Tic Tac Toe CLI
+
+Example of interacting with the move contract using the Sui Rust SDK
+
+## Quick start
+
+1, checkout the codebase
